Added benchmarks for mic.
Added the folder mic-benchmarks which has the same polybench bencmarks
with the offload pragma added to move work to the MIC card. Now only
added to adi (not tested for correctness) and for jacobi-1d.

Fixed a printing statment in the mic_access_sdk print_to_file (was
missing a \t)